Consolidate Gemini tools into single gemini_edit_tools.py file
- Combine four separate tool files into one consolidated module
- Update all imports to use the consolidated module
- Update tests to test individual tool functions
- Remove old separate tool files
- Maintain exact same functionality and tool signatures
- All tests still passing (4 unit tests, 6 integration tests, runtime tests)
