+++ openhands/agenthub/codeact_agent/function_calling.py
@@ -19,16 +19,10 @@ from openhands.agenthub.codeact_agent.tools import (
     create_cmd_run_tool,
     create_str_replace_editor_tool,
 )
-from openhands.agenthub.codeact_agent.tools.gemini_list_directory import (
+from openhands.agenthub.codeact_agent.tools.gemini_edit_tools import (
     create_gemini_list_directory_tool,
-)
-from openhands.agenthub.codeact_agent.tools.gemini_read_file import (
     create_gemini_read_file_tool,
-)
-from openhands.agenthub.codeact_agent.tools.gemini_replace import (
     create_gemini_replace_tool,
-)
-from openhands.agenthub.codeact_agent.tools.gemini_write_file import (
     create_gemini_write_file_tool,
 )
 from openhands.core.exceptions import (
